refactored into store/query
