+++ api/queries/neoQueries_test.go
@@ -1,14 +1,83 @@
-package store
+package queries
 
 import (
+	"context"
 	"fmt"
+	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j/dbtype"
 	"github.com/pennsieve/model-service-serverless/api/models"
+	"github.com/pennsieve/model-service-serverless/api/shared"
 	"github.com/stretchr/testify/assert"
+	"os"
 	"testing"
+	"time"
 )
 
-func TestQuery(t *testing.T) {
+var neo4jDriver neo4j.DriverWithContext
+
+func getEnv(key, fallback string) string {
+	if value, ok := os.LookupEnv(key); ok {
+		return value
+	}
+	return fallback
+}
+
+func TestMain(m *testing.M) {
+
+	// If testing on Jenkins (-> NEO4J_BOLT_URL is set) then wait for db to be active.
+	if _, ok := os.LookupEnv("NEO4J_BOLT_URL"); ok {
+		time.Sleep(30 * time.Second)
+	}
+
+	// Get Connection
+	testDBUri := getEnv("NEO4J_BOLT_URL", "bolt://localhost:7687")
+	testUserName := "neo4j"
+	testPassword := "blackandwhite"
+
+	var err error
+	neo4jDriver, err = neo4j.NewDriverWithContext(testDBUri,
+		neo4j.BasicAuth(testUserName, testPassword, ""),
+		func(config *neo4j.Config) {
+			config.MaxConnectionPoolSize = 10
+			config.MaxConnectionLifetime = 5 * time.Minute
+			config.ConnectionAcquisitionTimeout = 10 * time.Second
+		})
+	if err != nil {
+		panic(err)
+	}
+
+	// Seed NEO4J database
+
+	// Run tests
+	code := m.Run()
+
+	// return
+	os.Exit(code)
+}
+
+func TestQueries(t *testing.T) {
+	for scenario, fn := range map[string]func(
+		tt *testing.T, s *NeoQueries,
+	){
+		"create query syntax from query params": testCreateQuery,
+	} {
+		t.Run(scenario, func(t *testing.T) {
+			db := shared.NewNeo4jSession(neo4jDriver.NewSession(context.Background(), neo4j.SessionConfig{
+				AccessMode: neo4j.AccessModeWrite,
+			}))
+
+			queries := NewNeoQueries(db)
+
+			t.Cleanup(func() {
+				db.Close(context.Background())
+			})
+
+			fn(t, queries)
+		})
+	}
+}
+
+func testCreateQuery(t *testing.T, _ *NeoQueries) {
 
 	// Mock result of shortestPath method.
 	paths := []dbtype.Path{
