+++ lambda/service/handler/handler_test.go
@@ -1,16 +1,11 @@
-package service
+package handler
 
 import (
-	"context"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
-	"github.com/pennsieve/model-service-serverless/api/store"
-	"github.com/stretchr/testify/assert"
 	"os"
 	"testing"
 	"time"
 )
-
-var neo4jDriver neo4j.DriverWithContext
 
 func getEnv(key, fallback string) string {
 	if value, ok := os.LookupEnv(key); ok {
@@ -26,7 +21,7 @@ func TestMain(m *testing.M) {
 		time.Sleep(30 * time.Second)
 	}
 
-	// Get NEO4J URI based on local vs. dockerized testing
+	// Get Connection
 	testDBUri := getEnv("NEO4J_BOLT_URL", "bolt://localhost:7687")
 	testUserName := "neo4j"
 	testPassword := "blackandwhite"
@@ -50,29 +45,4 @@ func TestMain(m *testing.M) {
 
 	// return
 	os.Exit(code)
-}
-
-func TestDBConnection(t *testing.T) {
-
-	// Initiate NEO4j session
-	db := neo4jDriver.NewSession(context.Background(), neo4j.SessionConfig{
-		AccessMode: neo4j.AccessModeWrite,
-	})
-
-	t.Cleanup(func() {
-		db.Close(context.Background())
-	})
-
-	// Create GraphStore object with initiated db.
-	graphStore := store.NewGraphStore(db)
-	service := NewGraphService(graphStore)
-
-	err := service.store.InitOrgAndDataset(2, 1, "N:Organization:123", "N:Dataset:123")
-	assert.Nil(t, err, "Could not set organization and dataset")
-
-	models, err := service.GetDatasetModels(1, 2)
-	assert.Nil(t, err, "Could not get Dataset Models")
-
-	assert.Equal(t, 0, len(models))
-
 }
